chore: Migrate project from Java to Kotlin and remove lombok dependencies
BREAKING CHANGE: There is a good chance that this will introduce breaking changes although every effort has been put in to prevent any.

Some tests will remain in Java since most projects that rely on the lsd dependency will be in Java and we should make sure that we don't make it awkward or confusing to use.

Known issues:
* SynchronousResponse now inherits from SequenceMessage and not Message
* Only capture() methods have been made open for extension so some tests that mock the LsdContext may complain - switching to using spy(LsdContext.class) instead of mock(LsdContext.class) will fix some scenarios. If other methods need to be mocked they will need to be made open in a later commit
* Some builder classes have been replicated via kotlin utility classes but not all have bene preserved (e.g. if there was not really a need to have a builder because there weren't many arguments)