fix: #55 fix behaviour of hyperlink on arrows when in an iframe
Some tools like Jenkins are putting the report in an iframe which means the first click on an arrow (to view the data) is jumping out of the iframe (and losing the location the user was in when they clicked on the link). This change defaults the target value of the hyperlink to _self so that it shouldn't jump away from the iframe.